requirements.txt
Django>=4.0
channels>=4.0
channels-redis>=4.0
djangorestframework>=3.13
psycopg2-binary>=2.9
websockets>=10.0
pytest>=7.0
pytest-django>=4.5
